refactor: Split statistics into submodules
