+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.1'
 services:
   stats-server:
-    build: ./ewm-stats
+    build: ./ewm-stats/server
     image: ewm-stats
     container_name: ewm-stats
     ports:
