--- ewm-main/src/main/java/ru/practicum/ewm/client/StatsDtoRequest.java
@@ -1,13 +1,0 @@
-package ru.practicum.ewm.client;
-
-import lombok.Value;
-
-import java.time.LocalDateTime;
-
-@Value
-public class StatsDtoRequest {
-    String app;
-    String uri;
-    String ip;
-    LocalDateTime timestamp;
-}
