+++ ewm-main/src/main/java/ru/practicum/ewm/config/StatsClientConfig.java
@@ -1,0 +1,34 @@
+package ru.practicum.ewm.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import ru.practicum.stats.client.StatsClient;
+
+@Configuration
+@RequiredArgsConstructor
+public class StatsClientConfig {
+
+    @Value("${app.stats-server-url}")
+    private String serverUrl;
+
+    @Value("${app.name}")
+    private String app;
+
+    @Value("${app.format.date-time}")
+    private String format;
+
+    private final RestTemplateBuilder builder;
+
+    @Bean
+    public StatsClient makeStatsClient() {
+        return new StatsClient(
+                serverUrl,
+                app,
+                format,
+                builder
+        );
+    }
+}
