+++ ewm-stats/client/src/main/resources/application.yml
@@ -1,0 +1,5 @@
+app:
+  name: unknown
+  stats-server-url: http://localhost:9090
+  format:
+    date-time: yyyy-MM-dd HH:mm:ss
