+++ ewm-stats/dto/src/main/java/ru/practicum/stats/dto/StatsDtoRequest.java
@@ -1,0 +1,24 @@
+package ru.practicum.stats.dto;
+
+import lombok.Value;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+@Value
+public class StatsDtoRequest {
+    Long id;
+
+    @NotBlank
+    String app;
+
+    @NotBlank
+    String uri;
+
+    @NotBlank
+    String ip;
+
+    @NotNull
+    LocalDateTime timestamp;
+}
