+++ ewm-stats/dto/src/main/java/ru/practicum/stats/dto/StatsDtoResponse.java
@@ -1,0 +1,10 @@
+package ru.practicum.stats.dto;
+
+import lombok.Value;
+
+@Value
+public class StatsDtoResponse {
+    String app;
+    String uri;
+    Long hits;
+}
