+++ ewm-stats/server/src/main/java/ru/practicum/stats/server/StatsServer.java
@@ -1,0 +1,11 @@
+package ru.practicum.stats.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatsServer {
+    public static void main(String[] args) {
+        SpringApplication.run(StatsServer.class, args);
+    }
+}
