+++ ewm-stats/server/src/main/java/ru/practicum/stats/server/exeption/handler/ErrorResponse.java
@@ -1,0 +1,25 @@
+package ru.practicum.stats.server.exeption.handler;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Value;
+
+import java.time.OffsetDateTime;
+import java.util.List;
+
+@Value
+public class ErrorResponse {
+    @JsonProperty("offset-timestamp")
+    OffsetDateTime timestamp = OffsetDateTime.now();
+
+    @JsonProperty("end-point")
+    String endPoint;
+
+    @JsonProperty("status")
+    int status;
+
+    @JsonProperty("error")
+    String error;
+
+    @JsonProperty("reasons")
+    List<String> reasons;
+}
