+++ ewm-stats/server/src/main/java/ru/practicum/stats/server/stats/model/Stats.java
@@ -1,0 +1,38 @@
+package ru.practicum.stats.server.stats.model;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import lombok.experimental.Accessors;
+import lombok.experimental.FieldDefaults;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@ToString
+@Accessors(chain = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@Entity
+@Table(name = "stats")
+public class Stats {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "stats_id")
+    Long id;
+
+    @Column(name = "app", nullable = false)
+    String app;
+
+    @Column(name = "uri", nullable = false, length = 2047)
+    String uri;
+
+    @Column(name = "ip", nullable = false)
+    String ip;
+
+    @Column(name = "timestamp", nullable = false)
+    LocalDateTime timestamp;
+}
