+++ ewm-stats/server/src/main/java/ru/practicum/stats/server/stats/model/StatsMapper.java
@@ -1,0 +1,16 @@
+package ru.practicum.stats.server.stats.model;
+
+import org.springframework.stereotype.Component;
+import ru.practicum.stats.dto.StatsDtoRequest;
+
+@Component
+public class StatsMapper {
+    public Stats toEntity(StatsDtoRequest dto) {
+        return new Stats()
+                .setId(dto.getId())
+                .setApp(dto.getApp())
+                .setUri(dto.getUri())
+                .setIp(dto.getIp())
+                .setTimestamp(dto.getTimestamp());
+    }
+}
