+++ ewm-stats/server/src/main/resources/application-dev.yml
