+++ ewm-stats/server/src/main/resources/application-test.yml
