+++ ewm-stats/server/src/main/resources/application.yml
