+++ ewm-stats/server/src/main/resources/schema.sql
