--- ewm-stats/src/main/java/ru/practicum/ewm/stats/model/StatsDtoRequest.java
@@ -1,24 +1,0 @@
-package ru.practicum.ewm.stats.model;
-
-import lombok.Value;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
-
-@Value
-public class StatsDtoRequest {
-    Long id;
-
-    @NotBlank
-    String app;
-
-    @NotBlank
-    String uri;
-
-    @NotBlank
-    String ip;
-
-    @NotNull
-    LocalDateTime timestamp;
-}
