--- ewm-stats/src/main/java/ru/practicum/ewm/stats/model/StatsDtoResponse.java
@@ -1,10 +1,0 @@
-package ru.practicum.ewm.stats.model;
-
-import lombok.Value;
-
-@Value
-public class StatsDtoResponse {
-    String app;
-    String uri;
-    Long hits;
-}
